updateSentry

+++ config/RemovePages.js
@@ -5,7 +5,7 @@ fs.readdir(path.resolve(__dirname, '../pages'), (err, files) => {
 	if (err) throw err
 	if (files.includes('index.js')) {
 		files.forEach(file => {
-			if (file !== '404.js' && file !== '_app.js') {
+			if (file !== '404.js' && file !== '_app.js' && file !== '_error.js') {
 				fs.rm(
 					path.resolve(__dirname, '../pages', file),
 					{ recursive: true, force: true },
